backend/requirements.txt
Django==4.2.1
psycopg2==2.9.6
djangorestframework==3.14.0
django-webpack-loader==0.5.0
